% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{accum_input}
\alias{accum_input}
\title{Create an input file for Fortran program Partial.}
\usage{
accum_input(dem, k, d, length_scale, scratch_dir, out_file, overwrite = TRUE)
}
\arguments{
\item{dem}{The file name (full path) for the input DEM.}

\item{k}{Saturated hydraulic conductivity, in meters per hour.}

\item{d}{Storm duration in hours.}

\item{length_scale}{Diameter for smoothing the DEM.}

\item{scratch_dir}{Directory for temporary files. The input file is written
to the scratch_dir.}

\item{out_file}{File name (full path) for the output binary floating point
(.flt) raster.}
}
\value{
There is no explicit return object, but an explicit side effect
  is writing to disk of the partial input file.
}
\description{
Program Partial builds a raster giving the contributing area to each
DEM cell for a storm of specified duration. Inputs to Partial are read
from an ASCII input file using "Keyword: arguments" format.
}
