% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingData.R
\name{aggregatePolygonValues}
\alias{aggregatePolygonValues}
\title{Aggregate buffer values}
\usage{
aggregatePolygonValues(polygon, varName, fun)
}
\arguments{
\item{varName}{Name pattern of the variable to aggregate by}

\item{fun}{A function to aggregate buffer values by: max, min}

\item{values}{Values extracted from buffers, including an "ID" column}
}
\value{
A dataframe of buffer values.
}
\description{
Groups values by buffer and--for each buffer--keeps the entry
with the "fun" (min/max) variable value.
}
