% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_metrics.R
\name{align}
\alias{align}
\title{Align}
\usage{
align(
  refDTM = "nofile",
  alignDTM = "nofile",
  refDSM = "nofile",
  alignDSM = "nofile",
  iterations = 5,
  k = 0,
  dampener,
  outDTM = "nofile",
  tileNx = 0,
  tileNy = 0,
  overlap = 0.5,
  radius = 15,
  nslope = 7,
  maxSlope = 1,
  nAzimuth = 8,
  outbins = "nofile",
  outDif = "nofile",
  outOutlier = "nofile",
  scratch_dir = "none",
  executable_dir = "none",
  program_name = "align"
)
}
\arguments{
\item{refDTM}{Character: The file name (full path) for the input 
reference DTM.}

\item{alignDTM}{Character: The file name (full path) for the input
DTM to align with the reference DTM.}

\item{refDSM}{Character: The file name (full path) for the input
reference DSM (digital surface model).}

\item{alignDSM}{Character: The file name (full path) for the input
DSM to align with the reference DTM.}

\item{iterations}{Numeric (int): Number of iterations to solve for the shift}

\item{k}{Numeric (dbl): The number of inter-quartile ranges from q1 and q3
to use as a Tukey's fence for outlier removal}

\item{dampener}{Numeric (dbl): Dampener for the shift, 1.0 or less}

\item{outDTM}{Character: File name (full path) for the output aligned DTM}

\item{tileNx}{(int): number of tiles in the x direction}

\item{overlap}{(dbl): overlap between tiles}

\item{radius}{(dbl): radius in meters for measuring slope and aspect}

\item{nslope}{(int): number of gradient bins}

\item{maxSlope}{(dbl): maximum gradient for binning}

\item{nAzimuth}{(int): number of azimuth bins}

\item{outbins}{Character: output csv file of elevation differences
binned by slope and aspect}

\item{outDif}{Character: output elevation difference raster (.flt)}

\item{outOutlier}{Character: output outlier raster (.flt)}

\item{scratch_dir}{Character string: A scratch directory where temporary
files are written. If an input file for program partial is created,
it is written here.}

\item{executable_dir}{Character: The directory where the executable
file is located.}

\item{tilesNy}{(int): number of tiles in the y direction}
}
\value{
error code
}
\description{
Coregister two DTMs.
}
