% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{bldgrds_nochannels_input}
\alias{bldgrds_nochannels_input}
\title{Create an input file for Fortran program bldrds.}
\usage{
bldgrds_nochannels_input(
  dem,
  aspect_length,
  plan_length,
  grad_length,
  out_raster,
  scratch_dir
)
}
\arguments{
\item{dem}{Character string: The file name (full path) for the input DEM.}

\item{out_raster}{Character string: File name (full path) for the output
binary floating point (.flt) raster.}

\item{scratch_dir}{Character string: Directory for temporary files.
The input file is written to the scratch_dir.}

\item{aspect_length:}{double, length in meters over which aspect is measured.}

\item{plan_length:}{double, length in meters over which plan curvature is measured.}

\item{grad_length:}{double, length in meters over which gradient is measured.}
}
\value{
There is no explicit return object, but an explicit side effect
  is writing to disk of the input file.
}
\description{
Program bldgrds does flow routing for a DEM. This function
creates an input file specifying that the flow routing will
not involve delineation of channels. Therefore, flow directions
are entirely based on D-infinity.
}
