% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_models.R
\name{build_k_fold_rf_model}
\alias{build_k_fold_rf_model}
\title{Build a Random Forest model using cross-validation.}
\usage{
build_k_fold_rf_model(
  data = NULL,
  seed = 123,
  ctrl_method = "repeatedcv",
  k = 5,
  repeats = 3
)
}
\arguments{
\item{data}{The dataset to be used in model-building. Columns must contain
the predictor variables and a column called "class" that labels each row as
positive or negative.}

\item{seed}{A number setting the random seed to allow reproducible results.
Default is 123.}

\item{ctrl_method}{The method for}

\item{k}{The number of folds, or divisions of the data. Default is 5, which
produces a 80/20 ratio between training data and testing data for each
repetition.}

\item{repeats}{Number of times to repeat cross validation. The folds change
with each repeat.}
}
\value{
Nothing
}
\description{
This function acts as a wrapper for the train() function from
the \code{caret} package for building a classification Random Forest model
using the \code{randomForest} package.

The train function will compare models for a few values of tuning parameters.
For the randomForest model, the tuning parameter is \code{mtry}, which is the
number of variables randomly sampled as candidates at each split. For this
purpose, we use specificity, sensitivity, and area under the ROC curve (ROC)
as evaluating parameters.

\code{caret} generates a report which is printed, in addition to a plot
of the model performance across the tuning parameters and a report of the
time it took to generate the model.
}
