% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_models.R
\name{build_rf_model}
\alias{build_rf_model}
\title{Build Random Forest Model}
\usage{
build_rf_model(data = NULL, plot_model = FALSE)
}
\arguments{
\item{data}{A dataframe containing columns with model inputs and a
column called "class" which describes each row as "positive" or "negative"}

\item{plot_model}{Should a plot of the model be produced.}
}
\value{
A random forest model.
}
\description{
Build Random Forest Model
}
