% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTrainingPoints.R
\name{build_training_points}
\alias{build_training_points}
\title{Build training points from polygons}
\usage{
build_training_points(
  regionPolygon,
  wetlandPolys,
  wetlandSmpleRate,
  nonwetlandSampleRate,
  regionMargin,
  trainingPointsFile,
  wetlandTypes
)
}
\arguments{
\item{nonwetlandSampleRate}{Samples per km^2 of nonwetland}

\item{trainingPointsFile}{Name of training points file}

\item{wetlandTypes}{String: One or more of
\itemize{
  \item{Freshwater Forested/Shrub Wetland}
  \item{Freshwater Emergent Wetland}
  \item{Freshwater Pond}
  \item{Estuarine and Marine Wetland}
  \item{Riverine}
  \item{Lake}
  \item{Estuarine and Marine Deepwater}
  \item{Other}
}}

\item{regionPoly}{SpatVector polygon defining the region from which
points can be sampled}

\item{wetlandPolysFile}{Name of wetland polygons file}

\item{wetlandSampleRate}{Samples per km^2 of wetland}

\item{regionMrgin}{Width of region interior margin}
}
\description{
Build training points from polygons
}
