% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRandomForest.R
\name{buildRandomForest}
\alias{buildRandomForest}
\title{Build Random Forest}
\usage{
buildRandomForest(
  referenceRaster,
  inputRasterList,
  inputPolygonList = list(),
  trainingPoints,
  classFieldName,
  wetlandClass,
  nonwetlandClass,
  modelName,
  calcStats = FALSE,
  plotDir = NULL
)
}
\arguments{
\item{referenceRaster}{Raster to use as a grid reference}

\item{inputRasterList}{list of input rasters}

\item{inputPolygonList}{list of input polygons}

\item{trainingPoints}{Point feature classified by wetland type}

\item{classFieldName}{Name of the class field in the training dataset}

\item{wetlandClass}{Class name for wetlands}

\item{nonwetlandClass}{class name for non-wetlands}

\item{modelName}{Name for the random forest model}

\item{calcStats}{Calculate ROC statistics for the built model?}

\item{probRasterName}{Filename of the generated wetland probability raster}
}
\description{
Build Random Forest
}
