% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRocStats.R
\name{calcRocStats}
\alias{calcRocStats}
\title{Calculate ROC Statistics}
\usage{
calcRocStats(
  wetProbRaster = NULL,
  classifiedPoints = NULL,
  wetlandClass = NULL,
  nonWetlandClass = NULL
)
}
\arguments{
\item{wetProbRaster}{A \code{SpatRaster} object where each cell value
represents the probability of a wetland being there.}

\item{classifiedPoints}{A \code{??} object of points which are classified as
either "wetland" or "non-wetland" with the class names \code{wetlandClass}
and \code{nonWetlandClass}.}

\item{wetlandClass}{Name of the class indicating wetlands.}

\item{nonWetlandClass}{Name of the class indicating non-wetlands.}
}
\value{
A list of ROC statistic objects and values.
}
\description{
Calculates ROC statistics from a set of wetland-classified data
and a wetland probability raster. ROC statistics include: ROC, area under ROC
curve (AUC), precision, accuracy.
}
\examples{
\donttest{
library(TerrainWorksUtils)

wetProbRaster <- terra::rast("C:/Work/netmapdata/Puyallup/puy_prob.tif")

allPoints <- terra::vect("C:/Work/netmapdata/Puyallup/wetlandPnts.shp")
classifiedPoints <- allPoints[, "NEWCLASS"]

rocStats <- calcRocStats(wetProbRaster, classifiedPoints, "WET", "UPL")
}
}
