% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_metrics.R
\name{contributing_area}
\alias{contributing_area}
\title{Contributing area for a storm of fixed duration}
\usage{
contributing_area(
  input_file = "nofile",
  raster = "nofile",
  dem = "none",
  length_scale = 0,
  k = 0,
  d = 0,
  scratch_dir = "none"
)
}
\arguments{
\item{input_file}{Character string: a "partial" input file (optional).
If no input file is specified and no other arguments are present,
a Windows Explorer window opens for file selection.}

\item{raster}{Character string: file name (full path) for an existing
contributing-area raster to read from disk (optional).}

\item{dem}{Character string: The file name (full path) of the dem
(elevation raster) for construction of an input file.}

\item{length_scale}{Numeric (dbl): Length scale in meters over which to
smooth the DEM. This is the length used to measure gradient, curvature,
and guide flow directions.}

\item{k}{Numeric (dbl): Uniform saturated hydraulic conductivity for
construction of an input file.}

\item{d}{Numeric (dbl): Storm duration (dbl) in hours. Used for construction
of an input file for program partial.}

\item{scratch_dir}{Character string: A scratch directory where temporary
files are written. If an input file for program partial is created,
it is written here.}
}
\value{
A \code{SpatRaster} of contributing area for a storm of
  duration d hours.
}
\description{
Provide a \code{SpatRaster} giving the upslope contributing area to each
cell of a DEM for shallow subsurface groundwater flow to the cell.
D-infinity flow paths are traced upslope from each cell using a spatially
variable Darcy velocity (v = k*sin(gradient)) dependent on the specified
(spatially uniform) saturated hydraulic conductivity and the gradient of
each DEM cell.
}
\details{
Contributing_area operates in one of three modes, depending on the
input arguments:
\enumerate{
  \item As a wrapper for the Fortran "partial" executable,
  with an existing "partial" input_file.
  \item As a wrapper for program partial, but with the input file
  constructed by contributing_area.
  \item To read existing raster files from disk.
}
In modes 1 and 2, contributing_area calls program partial, which creates the
requested rasters and writes them to disk as floating point binary files.
These are then read and returned by contributing_area as a \code{SpatRaster}
object. In mode 3, existing raster files are read directly from disk and
returned as a \code{SpatRaster} object.
}
