% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_training_data.R
\name{create_training_data_with_buffer}
\alias{create_training_data_with_buffer}
\title{Create training data from a negative buffer region}
\usage{
create_training_data_with_buffer(
  positive_points,
  predictors_raster,
  pos_buffer = 15,
  neg_buffer = 50,
  analysis_region_mask = NULL,
  negative_proportion = 1,
  extraction_method = "centroid",
  extraction_layer = NULL
)
}
\arguments{
\item{positive_points}{SpatVector with locations of all points with positive
class}

\item{predictors_raster}{SpatRaster with a layer for each predictor variable}

\item{pos_buffer}{The radius around a positive point that should be
considered positive}

\item{neg_buffer}{The radius around a positive point that should be
considered negative}

\item{analysis_region_mask}{SpatRaster with non-NA values everywhere that
points can be sampled from. All locations that should be excluded from
sampling should be NA. If NULL, all cells which are non-NA for all layers
of the predictors_raster will be used.}

\item{negative_proportion}{Proportion of negative points to be generated
compared to number of positive points}

\item{extraction_method}{Method to use for extracting values from each point:
"all", "centroid", "max", or "min". Ignored if extractionPoints is not
polygon}

\item{extraction_layer}{Layer to use for extracting value. Ignored if
extraction_method = "centroid". Ignored if extraction_method is "all" or
"centroid" or if extractionPoints is not polygon.}
}
\description{
Given an input dataset and set of positive points, generate
randomly sampled negative points and extract predictor values for positive
and negative points. This sampling method is designed for a data set of
positive points that is not considered a full inventory, such as the DOGAMI
landslide inventory. First, we build a presumed negative region, with enough
proximity to recorded landslide initiation points that we assume if a
landslide had occurred there, it would have been recorded.
}
