% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_training_data.R
\name{create_training_points_from_polygons}
\alias{create_training_points_from_polygons}
\title{Sample training points from a set of polygons}
\usage{
create_training_points_from_polygons(
  polygons,
  analysis_region,
  sample_rate = 0.5,
  region_margin = 50,
  polygon_class = "positive",
  nonpolygon_class = "negative"
)
}
\arguments{
\item{polygons}{SpatVector of polygons indicating all locations belonging
to the class you wish to predict}

\item{analysis_region}{polygon or raster indicating the extent from which
points can be sampled. Only regions covered by non-NA cells will be included
if analysis_region is a raster.}

\item{sample_rate}{Samples per km^2}

\item{region_margin}{width in meters of margin to draw around polygon edges
which will not be used for sampling.}

\item{polygon_class}{label for points sampled from within polygons}

\item{nonpolygon_class}{label for points sampled outside polygons}
}
\description{
Sample training points from a set of polygons
}
