% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnalysisRegion.R
\name{createAnalysisRegionMask}
\alias{createAnalysisRegionMask}
\title{Create Analysis Region Mask}
\usage{
createAnalysisRegionMask(raster, points, maskVars = NULL, expansionFactor = 1)
}
\arguments{
\item{raster}{SpatRaster with layers for each analysis variable}

\item{points}{SpatVector of points or polygons to be used for calculating
range of allowed values for the analysis region}

\item{maskVars}{character vector of names of raster layers to include
in calculating allowable ranges. If NULL, all layers will be included.}

\item{expansionFactor}{Factor to expand the range by. Default is 1, which
indicates no expansion. 0.5 will reduce the range by 50%. 2 will double the
range.}
}
\value{
SpatRaster with NA values everywhere outside analysis region
and 1 everywhere within the analysis region.
}
\description{
Create a mask of the same size and shape as \code{raster} with
\code{NA} values everywhere outside of the analysis region. Cells included
in the anlaysis region are determined by the range of cell values covered
by \code{points}. First, the range of cell values which fall under
\code{points}  for each raster layer included in \code{maskVars} is
calculated, and only cells whose values fall within that range are included
in the analysis region.
The range can be expanded by a factor indicated by expansionFactor.
}
