% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingData.R
\name{createTrainingDataFromPoints}
\alias{createTrainingDataFromPoints}
\title{Create Training Data from Points}
\usage{
createTrainingDataFromPoints(
  positivePoints,
  predictorsRaster,
  analysisRegionMask = NULL,
  bufferRadius = 15,
  negativeProportion = 1,
  extractionMethod = "centroid",
  extractionLayer = NULL
)
}
\arguments{
\item{positivePoints}{SpatVector with locations of all points with positive
class}

\item{predictorsRaster}{SpatRaster with a layer for each predictor variable}

\item{analysisRegionMask}{SpatRaster with non-NA values everywhere that
points can be sampled from. All locations that should be excluded from
sampling should be NA. If NULL, all cells which are non-NA for all layers
of the predictorsRaster will be used.}

\item{bufferRadius}{minimum possible distance between a positive and negative
point}

\item{negativeProportion}{Proportion of negative points to be generated
compared to number of positive points}

\item{extractionMethod}{Method to use for extracting values from each point:
"all", "centroid", "max", or "min". Ignored if extractionPoints is not
polygon}

\item{extractionLayer}{Layer to use for extracting value. Ignored if
extractionMethod = "centroid". Ignored if extractionMethod is "all" or
"centroid" or if extractionPoints is not polygon.}
}
\value{
a data.frame with values for positive and negative points
}
\description{
Given an input dataset and set of positive points, generate
randomly sampled negative points and extract predictor values for positive
and negative points.
}
