% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingData.R
\name{createTrainingDataFromPolygons}
\alias{createTrainingDataFromPolygons}
\title{Create Training Data from Polygons}
\usage{
createTrainingDataFromPolygons(
  polygons,
  predictorsRaster,
  analysisRegion,
  sampleRate = 5,
  regionMargin = 50
)
}
\arguments{
\item{polygons}{SpatVector of polygons indicating all locations belonging
to the class you wish to predict}

\item{predictorsRaster}{SpatRaster with a layer for each predictor variable}

\item{analysisRegion}{polygon or raster indicating the extent from which
points can be sampled}

\item{sampleRate}{Samples per km^2}

\item{regionMargin}{width in meters of margin to draw around polygon edges
which will not be used for sampling.}
}
\description{
Given an input dataset and set of polygons indicating
a class, sample points from inside and outside polygons and extract
predictor values for positive and negative points.
}
