% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingData.R
\name{createTrainingPointsFromPolygons}
\alias{createTrainingPointsFromPolygons}
\title{Sample training points from a set of polygons}
\usage{
createTrainingPointsFromPolygons(
  polygons,
  analysisRegion,
  sampleRate = 0.5,
  regionMargin = 50,
  polygonClass = "positive",
  nonpolygonClass = "negative"
)
}
\arguments{
\item{polygons}{SpatVector of polygons indicating all locations belonging
to the class you wish to predict}

\item{analysisRegion}{polygon or raster indicating the extent from which
points can be sampled. Only regions covered by non-NA cells will be included
if analysisRegion is a raster.}

\item{sampleRate}{Samples per km^2}

\item{regionMargin}{width in meters of margin to draw around polygon edges
which will not be used for sampling.}

\item{polygonClass}{label for points sampled from within polygons}

\item{nonpolygonClass}{label for points sampled outside polygons}
}
\description{
Sample training points from a set of polygons
}
