% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_to_model.R
\name{dem_to_model}
\alias{dem_to_model}
\title{Dem to model}
\usage{
dem_to_model(
  dems,
  initiation_points,
  output_dir,
  elev_derivatives = c("plan", "norm", "mean", "tan"),
  pca_durations = c(5, 24),
  pca_conductivity = 1,
  length_scale = 15,
  use_analysis_mask = FALSE,
  analysis_mask_derivs = NULL,
  analysis_mask_expansion = 1,
  neg_region_buffer = 150,
  pos_region_buffer = 11,
  neg_sampling_proportion = 1,
  preprocess_norm = FALSE,
  preprocess_center = FALSE,
  plot_probability_raster = FALSE
)
}
\arguments{
\item{dems}{File path for the DEM of the study area.}

\item{initiation_points}{File path for the initiation points in the study area}

\item{output_dir}{A directory to put output files.}

\item{pca_durations}{The storm duration(s) in hours for the partial
contributing area calculation.}

\item{pca_conductivity}{The hydraulic conductivity used in the contributing
area calculation.}

\item{length_scale}{The length scale in m over which to calculate the
elevation derivatives.}

\item{use_analysis_mask}{Whether to use an analysis mask or not. This is an
optional step which limits the selection of negative data points to
areas on the DEM where the specific elevation derivatives fall within
the range observed in the initiation points.}

\item{analysis_mask_derivs}{Which elevation derivatives to use for
generating the analysis mask.}

\item{analysis_mask_expansion}{The expansion of the range to use for
generating the analysis mask.}

\item{neg_region_buffer}{The radius in m around the initiation points that
should be considered the negative region. We assume that any landslide
within this range would have been observed.}

\item{pos_region_buffer}{The radius in m around the initiation points that
should be considered part of the landslide initiation.}

\item{neg_sampling_proportion}{The proportion of negative to positive points
to be sampled from the negative region.}

\item{preprocess_norm}{Should training data be normalized before the model
is trained}

\item{preprocess_center}{Should training data be centered before the model
is trained}

\item{elev_derivates}{The elevation derivatives to calculate and include in
model generation. Current supported options are grad (gradient), prof
(profile curvature), plan (planar curvature), norm (normal slope
curvature), tan (tangential curvature), and mean (mean curvature).}
}
\value{
Returns the information needed to evaluate the model. For this, we
could probably look at just ROC, AUC.
}
\description{
Take a DEM and all additional parameters needed to build the
model. All parameters have default values that are based on a current
best-guess based on current literature or understanding of physical processes.
The purpose of this script is to make it easy to evaluate the sensitivity
of the model to various parameters.
}
