% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{DEV_input}
\alias{DEV_input}
\title{Create an input file for Fortran program LocalRelief.}
\usage{
DEV_input(dem, radius, out_raster, scratch_dir)
}
\arguments{
\item{dem}{Character string: The file name (full path) for the input DEM.}

\item{radius}{dbl: Radius in meters for calculating DEV.}

\item{out_raster}{Character string: File name (full path) for the output
binary floating point (.flt) raster.}

\item{scratch_dir}{Character string: Directory for temporary files.
The input file is written to the scratch_dir.}
}
\value{
There is no explicit return object, but an explicit side effect
  is writing to disk of the input file.
}
\description{
Program LocalRelief builds a deviation-from-local-elevation (DEV) raster
(and other things as well, DEV is all we need here).
}
