% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{distanceToRoad_input}
\alias{distanceToRoad_input}
\title{Create an input file for Fortran program distanceToRoad.}
\usage{
distanceToRoad_input(dem, radius = 1000, road_file, out_raster, scratch_dir)
}
\arguments{
\item{dem}{Character string: The file name (full path) for the input DEM.}

\item{radius}{Double: radius in meters to extend from a road.}

\item{road_file}{Character string: input road polyline shapefile.}

\item{out_raster}{Character string: File name (full path) for the output
binary floating point (.flt) raster.}

\item{scratch_dir}{Character string: Directory for temporary files.
The input file is written to the scratch_dir.}
}
\value{
There is no explicit return object, but an explicit side effect
  is writing to disk of the input file.
}
\description{
Program distanceToRoad builds a raster giving the distance in meters
to the nearest road for each DEM grid point.
Inputs to distanceToRoad are read from an ASCII input file
using "Keyword: arguments" format.
}
