% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis_region.R
\name{extract_range}
\alias{extract_range}
\title{Extract range of values from raster}
\usage{
extract_range(raster, extraction_locations, expansion_factor = 1)
}
\arguments{
\item{raster}{A SpatRaster of explanatory variables}

\item{extraction_locations}{A SpatVector of points or polyogons to extract range from.}

\item{expansion_factor}{Factor to expand the range by. 1 indicates no expansion.
0.5 will reduce the range by 50\%. 2 will double the range.}
}
\value{
A matrix that holds the min & max initiation limits of each raster
layer
}
\description{
Finds the range of values for each layer of a raster which
fall under a set of points or polygons.
}
