% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_training_data.R
\name{extract_values}
\alias{extract_values}
\title{Extract buffer values}
\usage{
extract_values(
  raster,
  points,
  extraction_method = "all",
  extraction_layer = NULL,
  return_type = "table",
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{raster}{A SpatRaster of explanatory variables with a layer for
each variable to extract.}

\item{points}{SpatVector of points to use for extraction. If
polygons, the point chosen for extraction will be determined via the
extraction_method parameter.}

\item{extraction_method}{Method to use for extracting values from each point:
"all", "centroid", "max", or "min". Ignored if extractionPoints is not
polygon}

\item{extraction_layer}{Layer to use for extracting value. Ignored if
extraction_method = "centroid". Ignored if extraction_method is "all" or
"centroid" or if extractionPoints is not polygon.}

\item{return_type}{\code{"table"} or \code{"points"}. If \code{"table"}
(default) a data.table with values is returned. if \code{"points"}, a
\code{SpatVector} of points is returned, with training values as attributes
for each point.}

\item{na.rm}{Remove any cells with NA values?}

\item{...}{Additional arguments passed onto \code{terra::extrac}}
}
\value{
A dataframe of extracted raster values with an additional "class"
column.
}
\description{
Extracts values for all layers of a raster for a set of points
or polygons. If
}
