% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingData.R
\name{extractBufferValues}
\alias{extractBufferValues}
\title{Extract buffer values}
\usage{
extractBufferValues(
  raster,
  positiveBuffers,
  negativeBuffers,
  extractionMethod = "all",
  extractionLayer = NULL
)
}
\arguments{
\item{raster}{A SpatRaster of explanatory variables}

\item{positiveBuffers}{A SpatVector of buffers for each positive}

\item{negativeBuffers}{A SpatVector of buffers for each negative}

\item{extractionMethod}{Method to use for extracting values from each buffer:
"all", "center", "max", or "min"}

\item{extractionLayer}{Layer to use for extracting value. Ignored if
extractionMethod = "center"}
}
\value{
A dataframe of extracted raster values with an additional "class"
column.
}
\description{
Extracts variable raster values within initiation and
non-initiation buffers.
}
