% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRasterValues.R
\name{extractRasterValues}
\alias{extractRasterValues}
\title{Extract raster values}
\usage{
extractRasterValues(raster = NULL, points = NULL)
}
\arguments{
\item{raster}{A \code{SpatRaster} object to extract values from.}

\item{points}{A \code{SpatVector} object of points.}
}
\value{
A \code{data.frame} of values.
}
\description{
Extracts all raster variable values at specified locations.
}
\examples{
\donttest{
library(TerrainWorksUtils)

points <- terra::vect("C:/Work/netmapdata/pack_forest/PF_trainingdata.shp")

# Single-band rasters
continuousRaster <- terra::rast("C:/Work/netmapdata/pack_forest/pf_dtm3.flt")
factorRaster <- terra::rast("C:/Work/netmapdata/pack_forest/geo_unit.tif")
factorRaster <- fixFactorRaster(factorRaster)

v1 <- extractRasterValues(continuousRaster, points)
v2 <- extractRasterValues(factorRaster, points)

# Multi-band raster
alignedRasters <- alignRasters(continuousRaster, list(continuousRaster, factorRaster))
rasterStack <- c(alignedRasters[[1]], alignedRasters[[2]])

v3 <- extractRasterValues(rasterStack, points)
}
}
