% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingData.R
\name{extractValues}
\alias{extractValues}
\title{Extract buffer values}
\usage{
extractValues(
  raster,
  points,
  extractionMethod = "all",
  extractionLayer = NULL,
  xy = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{raster}{A SpatRaster of explanatory variables with a layer for
each variable to extract.}

\item{points}{SpatVector of points to use for extraction. If
polygons, the point chosen for extraction will be determined via the
extractionMethod parameter.}

\item{extractionMethod}{Method to use for extracting values from each point:
"all", "centroid", "max", or "min". Ignored if extractionPoints is not
polygon}

\item{extractionLayer}{Layer to use for extracting value. Ignored if
extractionMethod = "centroid". Ignored if extractionMethod is "all" or
"centroid" or if extractionPoints is not polygon.}

\item{xy}{Return coordinates of cells?}

\item{na.rm}{Remove any cells with NA values?}
}
\value{
A dataframe of extracted raster values with an additional "class"
column.
}
\description{
Extracts values for all layers of a raster for a set of points
or polygons. If
}
