% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorizeRaster.R
\name{factorizeRaster}
\alias{factorizeRaster}
\title{Properly format a factor raster}
\usage{
factorizeRaster(raster)
}
\description{
Properly format a factor raster
}
\details{
terra seems to have problems reading values from some factor raster
files. A factor raster made using the ArcGIS 'Polygon to Raster' tool will
only show a 'Count' field when loaded in terra, and even that seems to be
misleveled by 1 row when you look at the terra::cats() table for the raster.
}
