% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorizeRaster.R
\name{fixFactorRaster}
\alias{fixFactorRaster}
\title{Fix a misfactored raster}
\usage{
fixFactorRaster(raster = NULL)
}
\arguments{
\item{raster}{The faulty factor \code{SpatRast} object.}
}
\description{
Creates a correctly-factored version of a misfactored
\code{SpatRast} object.
}
\details{
terra seems to have issues loading some factor raster files. For
instance, a factor raster made using the ArcGIS 'Polygon to Raster' tool and
then loaded with \code{terra::rast()} will only show a 'Count' field instead
of the field specified in 'Polygon to Raster'. Additionally, the 'Count'
value assigned to each cell appears to be misleveled by 1 row when inspected
in the raster's \code{terra::cats()} table. The \code{fixFactorRaster()}
function attempts to take a faulty raster and map its values to the correct
field, as it still exists in the \code{terra::cats()} table.
}
\examples{
\donttest{
library(TerrainWorksUtils)

faultyRaster <- terra::rast("C:/Work/netmapdata/pack_forest/geo_unit.tif")
faultyRaster
terra::cats(faultyRaster)

fixedRaster <- fixFactorRaster(faultyRaster)
fixedRaster
terra::cats(fixedRaster)
}
}
