% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_to_model.R
\name{generate_success_curve}
\alias{generate_success_curve}
\title{Generate success curve}
\usage{
generate_success_curve(
  predicted_raster,
  initiation_points,
  step_size = 0.1,
  silence_plot = FALSE
)
}
\arguments{
\item{predicted_raster}{A raster of the study area with predicted
probabilities. This must have a layer called "positive", indicating the
probability of an initiation point, with values ranging from 0 to 1.}

\item{initiation_points}{A vector object with initiation points.}

\item{step_size}{A parameter indicating how big each probability region range
should be.}

\item{silence_plot}{If TRUE, the plot will not be produced.}
}
\value{
The values associated with the success curve.
}
\description{
Produces a success curve for a give prediction raster. The
success curve divides the predicted raster into 10 probability regions, then
finds the proportion of total inititation points in each probability region.
}
