% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{get_args}
\alias{get_args}
\title{Get a character vector of arguments from an input-file line.}
\usage{
get_args(infile, line_num)
}
\arguments{
\item{infile;}{a tibble created by get_input_file.}

\item{line_num;}{the line of the input file.}
}
\value{
arguments; a character vector of arguments
}
\description{
Get a character vector of arguments from an input-file line.
}
