% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{get_dem}
\alias{get_dem}
\title{Get a list of DEM raster files listed in an ASCII input file.}
\usage{
get_dem(infile)
}
\arguments{
\item{infile, }{a tibble created by get_input_file.}
}
\value{
A list of character strings, each string gives a DEM file name
}
\description{
Get a list of DEM raster files listed in an ASCII input file.
}
