% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceMetrics_utils.R
\name{get_executable_path}
\alias{get_executable_path}
\title{Get executable path}
\usage{
get_executable_path()
}
\value{
The directory of the executable files.
}
\description{
Returns the filepath to the executable files for calculating elevation
derivatives. There must be a .zip file called "files.zip" in a directory
called "DEMutilities", OR the executables must be unzipped into a
"/DEMutilities/files/" folder.
}
