% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{get_keyword}
\alias{get_keyword}
\title{Get the keyword from an input-file line.}
\usage{
get_keyword(infile, line_num)
}
\arguments{
\item{infile, }{a tibble created by get_input_file.}

\item{line_num, }{line number in the input file.}
}
\value{
A character vector containing the keyword;
  NA if no keyword is present.
}
\description{
Get the keyword from an input-file line.
}
\examples{
* infile <- get_input_file() select input file interactively.
* keyword <- get_keyword(infile,3) gets the keyword from
  line 3 of the selected file.

}
