% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_metrics.R
\name{huntLS}
\alias{huntLS}
\title{HuntLS
Program HuntLS reads an "outlier" raster created by program Align.
The "outlier" k values indicate the number of interquartile ranges
(q3-q1) a DoD de (delta elevation) value falls outside of the 
interquartile range: k = (de-q1)/(q3-a1) if k is less than q1 
and k = (de-q3)/(q3-q1) if k is greater than q3. Values of de
between q1 and q3 are given k values of zero. HuntLS associates 
k patches with field-mapped landslide locations surveyed for the
Post Mortem (Stewart et al., 2013) study.}
\usage{
huntLS(
  DEM = "nofile",
  Outlier = "nofile",
  DoD = "nofile",
  Accum = "nofile",
  AccumThreshold = -9999,
  LSpnts = "nofile",
  IDfield = "none",
  Radius = -9999,
  AspectLength = -9999,
  GradLength = -9999,
  OutlierThreshold = -9999,
  ScratchDir = "nofile",
  OutPatch = "nofile",
  OutGrad = "nofile",
  Outcsv = "nofile",
  executable_dir = "none"
)
}
\arguments{
\item{DEM}{Character: The reference DTM used for program Align.}

\item{Outlier}{Character: The input outlier raster (.flt) created by Align.}

\item{DoD}{Character: Input DoD raster (.flt) created by Align.}

\item{Accum}{Character: Input flow accumulation raster created by bldgrds.}

\item{AccumThreshold}{Numeric (dbl): Contributing area above which landslide
patches are precluded.}

\item{LSpnts}{Character: Input point shapefile of mapped landslide points.}

\item{IDfield}{Character: LSpnts attributes-table field name for record ID.}

\item{Radius}{Numeric (dbl): Search radius in meters for matching landslide
points to k-value patches.}

\item{AspectLength}{Numeric (dbl): Length in meters to calculate aspect.}

\item{GradLength}{Numeric (dbl): Length in meters to calculate gradient.}

\item{OutlierThreshold}{Numeric (dbl): Minimum absolute k value for a patch.}

\item{ScratchDir}{Character: Scratch directory.}

\item{OutPatch}{Character: Output patch raster (.flt).}

\item{OutGrad}{Character: Output gradient raster (.flt).}

\item{Outcsv}{Character: Output comma-delimited table.}

\item{executable_dir}{Character: The directory where the executable
file is located.}
}
\value{
returnCode, a value of zero indicates success
}
\description{
HuntLS
Program HuntLS reads an "outlier" raster created by program Align.
The "outlier" k values indicate the number of interquartile ranges
(q3-q1) a DoD de (delta elevation) value falls outside of the 
interquartile range: k = (de-q1)/(q3-a1) if k is less than q1 
and k = (de-q3)/(q3-q1) if k is greater than q3. Values of de
between q1 and q3 are given k values of zero. HuntLS associates 
k patches with field-mapped landslide locations surveyed for the
Post Mortem (Stewart et al., 2013) study.
}
