% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_metrics.R
\name{LS_poly}
\alias{LS_poly}
\title{LS_poly}
\usage{
LS_poly(
  DEM = "nofile",
  polyFile = "nofile",
  polyID = "none",
  inGrad = "nofile",
  inTan = "nofile",
  inProf = "nofile",
  inFoS = "nofile",
  outGrad = "nofile",
  gradRadius = -9999,
  outTan = "nofile",
  tanRadius = -9999,
  outProf = "nofile",
  profRadius = -9999,
  outNodes = "nofile",
  outCsv = "nofile",
  outInit = "nofile",
  scratchDir = "none",
  executableDir = "none"
)
}
\arguments{
\item{DEM}{Character: input DEM (.flt or .tif), full path name}

\item{polyFile}{Character: input landslide polygon shapefile}

\item{polyID}{Character: name of the ID field for the input polygons}

\item{inGrad}{Character: input gradient raster (.flt or .tif), full path name}

\item{inTan}{Character: input tangential curvature raster (.flt or .tif), full path name}

\item{inProf}{Character: input profile curvature raster (.flt or .tif), full path name}

\item{inFoS}{Character: input factor of safety (FoS) raster (.flt or .tif), full path name}

\item{outGrad}{Character: output gradient raster (.flt or .tif), full path name}

\item{gradRadius}{Numeric (dbl): radius in meters for calculating gradient}

\item{outTan}{Character: output tangential curvature raster (.flt or .tif), full path name}

\item{tanRadius}{Numeric (dbl): radius in meters for calculating tangential curvature}

\item{outProf}{Character: output profile curvature raster (.flt or .tif), full path name}

\item{profRadius}{Numeric (dbl): radius in meters for calculating profile curvature}

\item{outNodes}{Character: output node point shapefile (.shp), full path name}

\item{outCsv}{Character: output csv file with patch statistics, full path name}

\item{outInit}{Character: output initiation zone raster (.flt or .tif), full path name}

\item{scratchDir}{Character: scratch directory, full path name}
}
\value{
returnCode, a value of zero indicates success
}
\description{
LS_poly reads a polygon shapefile of landslide scars. 
It generates centerlines through each polygon and builds a 
linked-node list through the centerlines. A likely initiation zone
is delineated from the upslope end of each polygon, extending downslope
from the top of the polygon a distance equal to the average polygon width.
Statistics are calculated for the initiation zones, including the metrics
for gradient, tangential curvature, profile curvature, and Factor-of-Safety
(FoS). Gradient and curvature rasters can be read from disk or calculated.
If calculated by LS_poly, radius values must be provided. The calculated
rasters may be written to disk for use on subsequent runs of LS_poly.
A FoS input raster must be specified. This may be generated with the 
FoS program.
}
