% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{LS_poly_input}
\alias{LS_poly_input}
\title{Create an input file for Fortran program LS_poly}
\usage{
LS_poly_input(
  DEM,
  polyFile,
  polyID,
  inGrad = "nofile",
  inTan = "nofile",
  inProf = "nofile",
  inFoS,
  outGrad = "nofile",
  gradRadius = 0,
  outTan = "nofile",
  tanRadius = 0,
  outProf = "nofile",
  profRadius = 0,
  outNodes,
  outCsv,
  outInit,
  scratchDir
)
}
\arguments{
\item{DEM}{Character: input DEM (.flt or .tif), full path name}

\item{polyFile}{Character: input landslide polygon shapefile}

\item{polyID}{Character: name of the ID field for the input polygons}

\item{inGrad}{Character: input gradient raster (.flt or .tif), full path name}

\item{inTan}{Character: input tangential curvature raster (.flt or .tif), full path name}

\item{inProf}{Character: input profile curvature raster (.flt or .tif), full path name}

\item{inFoS}{Character: input factor of safety (FoS) raster (.flt or .tif), full path name}

\item{outGrad}{Character: output gradient raster (.flt or .tif), full path name}

\item{gradRadius}{Numeric (dbl): radius in meters for calculating gradient}

\item{outTan}{Character: output tangential curvature raster (.flt or .tif), full path name}

\item{tanRadius}{Numeric (dbl): radius in meters for calculating tangential curvature}

\item{outProf}{Character: output profile curvature raster (.flt or .tif), full path name}

\item{profRadius}{Numeric (dbl): radius in meters for calculating profile curvature}

\item{outNodes}{Character: output node point shapefile (.shp), full path name}

\item{outCsv}{Character: output csv file with patch statistics, full path name}

\item{outInit}{Character: output initiation zone raster (.flt or .tif), full path name}

\item{scratchDir}{Character: scratch directory, full path name}
}
\value{
returnCode, a value of zero indicates success
}
\description{
Create an input file for Fortran program LS_poly
}
