% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_metrics.R
\name{LShunter}
\alias{LShunter}
\title{LShunter}
\usage{
LShunter(
  DoD = "nofile",
  Outlier = "nofile",
  threshold1 = -9999,
  threshold2 = -9999,
  Gradient = "nofile",
  min1 = -9999,
  min2 = -9999,
  Accum = "nofile",
  maxAccum1 = -9999,
  maxAccum2 = -9999,
  minSize = -9999,
  OutPatch = "nofile",
  ScratchDir = "none",
  Executable_dir = "none"
)
}
\arguments{
\item{DoD}{Character: The reference DoD (DEM of Difference) raster (.flt) created by Align.}

\item{Outlier}{Character: The outlier raster (.flt) created by Align.}

\item{threshold1}{Numeric (dbl): Outlier threshold for the 1st round.}

\item{threshold2}{Numeric (dbl): Outlier threshold for the 2nd round.}

\item{Gradient}{Character: An input gradient raster (.flt) created by program HuntLS.}

\item{min1}{Numeric (dbl): Minimum gradient for the 1st round.}

\item{min2}{Numeric (dbl): Minimum gradient for the 2nd round.}

\item{Accum}{Character: Input flow accumulation raster (.flt) created by bldgrds.}

\item{maxAccum1}{Numeric (dbl): Maximum flow accumulation for the 1st round.}

\item{maxAccum2}{Numeric (dbl): Maximum flow accumulation for the 2nd round.}

\item{OutPatch}{Character: Output patch raster (.flt)}

\item{ScratchDir}{Character: Scratch directory.}

\item{Executable_dir}{Character: The directory where the executable file is located}

\item{MinSize}{Numeric (dbl): Minimum size of a patch in square meters.}
}
\value{
returnCode, a value of zero indicates success
}
\description{
Program LShunter identifies potential landslide sites using 
"outlier patches" on a DoD (DEM of Difference) delineated on
an "outlier" raster created by Program Align.
}
