% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_negative_points.R
\name{make_neg_region}
\alias{make_neg_region}
\title{Make a negative region}
\usage{
make_neg_region(initiation_points, inner_buffer = 5, outer_buffer = 25)
}
\arguments{
\item{initiation_points}{A SpatVector that contains the initiation points.}

\item{inner_buffer_rad}{The size of the inner buffer, as a radius. This is
the area around the initiation point that should be considered "positive",
or part of the landslide.}

\item{outer_buffer_rad}{The size of the outer buffer, as a radius. This is
the area around the landslide that can be considered "negative", or if there
had been a landslide here, it would have been recorded.}
}
\value{
SpatVector with polygons outlining the presumed negative region.
}
\description{
Creates a negative region with which to sample from that exists
between an inner buffer and an outer buffer around given initiation points.
This method will be used to sample a landslide database that is not
considered a full inventory, i.e., we cannot assume all locations that are
not marked as landslide initiation points did not have landslide.
}
