% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_to_model.R
\name{make_prob_raster}
\alias{make_prob_raster}
\title{Make probability raster}
\usage{
make_prob_raster(model, data, plot = FALSE)
}
\arguments{
\item{model}{A model object that can be called with the predict function.}

\item{data}{The full dataset to be predicted.}

\item{plot}{Whether to produce a plot of the prediction raster.}
}
\value{
The predicted raster
}
\description{
Make probability raster
}
