% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{makegrids_input}
\alias{makegrids_input}
\title{Create an input file for Fortran program makegrids}
\usage{
makegrids_input(dem, length_scale, scratch_dir, rasters, overwrite = TRUE)
}
\arguments{
\item{dem:}{File name (full path) to input DEM.}

\item{length_scale:}{Diameter in meters over which to measure attributes.}

\item{scratch_dir:}{Directory for storing temporary files. The input
file is written to the scratch_dir.}

\item{rasters:}{A character vector. Each element contains two strings
separated by a comma. The first specifies the type of attribute,
the second specifies the file name (full path) for the output floating
point binary (.flt) raster.}

\item{overwrite:}{If TRUE, allows overwriting of an existing input file.}
}
\value{
There is no return object, but an explicit side effect is
  writing to disk of the makegrids input file.
}
\description{
MakeGrids is called by functions in surface_metrics.
MakeGrids reads an ASCII input file with a "Keyword: arguments" format.
}
