% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis_region.R
\name{mask_by_range}
\alias{mask_by_range}
\title{Mask a raster}
\usage{
mask_by_range(raster, range_mx)
}
\arguments{
\item{raster}{A SpatRaster}

\item{range_mx}{A matrix with one row for each layer of the raster to
use when calculating mask, and a column each for "min" and "max", indicating
min and max allowable values. Rownames must match layer names in raster.}
}
\value{
A SpatRaster mask: Values are NA where values fall outside the
range and 1 where values fall within the range.
}
\description{
Create a raster mask based on an allowable range of
values
}
