% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_metrics.R
\name{modelDensity}
\alias{modelDensity}
\title{modelDensity}
\usage{
modelDensity(
  mask = "nofile",
  init_pnts = "nofile",
  intercept = -1e+30,
  R4rasters,
  I4rasters,
  unit = "unknown",
  prop_raster = "nofile",
  density_raster = "nofile",
  ROC = "nofile",
  scratch_dir = "none",
  executable_dir = "none"
)
}
\arguments{
\item{mask}{Character: Input mask raster (.flt)}

\item{init_pnts}{Character: initiation points (.shp)}

\item{intercept}{Numeric (dpl): Model intercept}

\item{R4rasters}{list: Continuous covariate rasters. 
Name, file, number of coefficients, list of coefficient values}

\item{I4rasters}{list: Integer factor covariate rasters.
Name, file, number of classes, minimum class, maximum class,
list of class values and associated coefficient value}

\item{unit}{Character: "KM" to convert units to kilometers}

\item{prop_raster}{Character: output proportion raster (.flt)}

\item{density_raster}{Character: Output density raster (.flt)}

\item{ROC}{Character: Output ROC csv file}

\item{scratch_dir}{Character: Folder to use as scratch space on disk}

\item{executable_dir}{Character: Folder where executable file is located}
}
\value{
returnCode, a value of zero indicates success
}
\description{
modelDensity builds a density raster for a Poisson Point model.
The model covariate rasters and coefficients are specified in this 
input file, along with the points used to build the model.
ModelDensity also provides:
- an ROC curve csv file using the modeled density and input points
- a proportion-of-points raster calculated from the density raster
- a csv file showing the actual proportion of points in 
  each 10% increment of the proportion raster
}
