% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{parse_arg}
\alias{parse_arg}
\title{Parse "parameter = value" pairs from a specified element
  in a list of arguments.}
\usage{
parse_arg(arguments, i = 1)
}
\arguments{
\item{arguments;}{a character vector of arguments output by get_args.}

\item{i;}{the element in the argument vector to parse.}
}
\value{
A 2-element character vector. The first element gives the
  parameter name; the second element gives the parameter value.
  If no equal sign is present in the argument, the parameter element
  is empty and the argument value is contained in the value element.
}
\description{
Parse "parameter = value" pairs from a specified element
  in a list of arguments.
}
