% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_metrics.R
\name{PFA_debris_flow}
\alias{PFA_debris_flow}
\title{Function PFA_debris_flow;
terrain attributes along debris-flow runout path.
A wrapper for Fortran program PFA_debris_flow.}
\usage{
PFA_debris_flow(
  dem = "none",
  init_points = "nofile",
  geo_poly = "nofile",
  stand_age = "nofile",
  tracks = "nofile",
  radius = 0,
  initRadius = 0,
  length_scale = 0,
  slope_intercept = 0,
  slope_coef = 0,
  bulk_coef = 0,
  init_width = 0,
  init_length = 0,
  DF_width = 0,
  alpha = 0,
  uncensored,
  scratch_dir = "none"
)
}
\arguments{
\item{dem}{Character string: The file name (full path) of the dem
(elevation raster) for construction of an input file.}

\item{radius}{double: Distance in meters to extend the search for a road.}

\item{uncensored}{string: if "TRUE", interpret all end-point tracks as uncensored, even without mapped deposition}

\item{scratch_dir}{Character string: A scratch directory where temporary
files are written. If an input file for program partial is created,
it is written here.}

\item{input_file}{Character string: a "partial" input file (optional).
If no input file is specified and no other arguments are present,
a Windows Explorer window opens for file selection.}

\item{init_points:}{string: File name for initiation-point shapefile.}

\item{geo_poly:}{string: File name for rock-type polygon shapefile.}

\item{stand_age:}{string: File name for LEMMA stand-age flt raster.}

\item{tracks:}{string: File name for DOGAMI debris-flow-track polyline shapefile.}

\item{radius:}{dbl: Search radius in meters for matching DEM flow path to DOGAMI track.}

\item{length_scale:}{dbl: Length in meters to measure elevation derivatives.}

\item{bulk_coef:}{dbl: Coefficient for linear bulking equation}

\item{alpha:}{dbl: Proportion of debris-flow cross-sectional volume deposited per unit length.}
}
\value{
A csv data file.
}
\description{
Provides a csv file that can be read to a dataframe for a Cox survival model.
}
\details{
A list of multinomial-logistic-regression-model coefficients for probability
of scour and deposition are required; these are hard wired into this code.
}
