% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{PFA_debris_flow_input}
\alias{PFA_debris_flow_input}
\title{Function PFA_debris_flow_input;
create an input file for Fortran program PFA_debris_flow.}
\usage{
PFA_debris_flow_input(
  dem,
  init_points,
  geo_poly,
  stand_age,
  tracks,
  radius,
  initRadius,
  length_scale,
  slope_intercept,
  slope_coef,
  bulk_coef,
  init_width,
  init_length,
  DF_width,
  alpha,
  uncensored,
  scratch_dir,
  out_surv,
  out_point,
  out_kaplanMeier,
  coef
)
}
\arguments{
\item{dem}{Character string: The file name (full path) for the input DEM.}

\item{scratch_dir}{Character string: Directory for temporary files.
The input file is written to the scratch_dir.}

\item{init_points:}{string: File name for initiation-point shapefile.}

\item{geo_poly:}{string: File name for rock-type polygon shapefile.}

\item{stand_age:}{string: File name for LEMMA stand-age flt raster.}

\item{tracks:}{string: File name for DOGAMI debris-flow-track polyline shapefile.}

\item{radius:}{dbl: Search radius in meters for matching DEM flow path to DOGAMI track.}

\item{length_scale:}{dbl: Length in meters to measure elevation derivatives.}

\item{bulk_coef:}{dbl: Bulking coefficient}

\item{alpha:}{dbl: Proportion of debris-flow cross-sectional volume deposited per unit length.}

\item{uncensored:}{string: If TRUE, interpret all endpoint tracks as uncensored.}

\item{out_track:}{string: Output track point shapefile.}

\item{out_surv:}{string: Output input file for Cox survival.}

\item{out_point:}{string: Output initiation point shapefile; includes fields for scour and deposit volume.}

\item{coef:}{list of dbl values specifying multinomial logistic regression coefficients
for probability of scour, deposition, and transitional flow.}
}
\value{
There is no explicit return object, but an explicit side effect
  is writing to disk of the input file.
}
\description{
PFA_debris_flow is one of a sequence of programs used to assemble data files
for the PFA recalibration of the landslide initiation and debris-flow-runout models.
The Quatro file PFA_runout describes calibration of a multinomial model for
estimating the probability of debris-flow scour or deposition using the ODF
1996 Storm Study surveys. The coefficients for that model are inputs to PFA_debris_flow.
Here, the debris-flow runout tracks from the DOGAMI Special Paper 53 study are
associated with lidar-DEM flow paths and terrain attributes associated with runout
extent are assembled and output in a format for a Cox survival model with
time-dependent covariates.
}
