% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_metrics.R
\name{quantiles}
\alias{quantiles}
\title{Quantiles}
\usage{
quantiles(
  in_raster = "nofile",
  radius = 0,
  buffer = 0,
  out_outlier = "nofile",
  out_q1 = "nofile",
  out_q2 = "nofile",
  out_q3 = "nofile",
  out_mean = "nofile",
  out_zscore = "nofile",
  out_prob = "nofile",
  scratch_dir = "none",
  executable_dir = "none",
  program_name = "quantiles"
)
}
\arguments{
\item{in_raster}{Character: The file name (full path) for the input raster.}

\item{radius}{Numeric (dbl): radius in meters for the moving window}

\item{buffer}{Numeric (dbl): spacign between moving window center points,
in raster cells.}

\item{out_outlier}{Character: output file name (full path) for the 
outlier raster (optional)}

\item{out_q1}{Character: output file name (full path) for the
first-quartile raster (optional)}

\item{out_q2}{Character: output file name (full path) for the
median raster (optional)}

\item{out_q3}{Character: output file name (full path) for the
third-quartile raster (optional)}

\item{out_mean}{Character: output file name (full path) for the mean raster}

\item{out_zscore}{Character: output file name (full path) for the 
z-score raster}

\item{out_prob}{Character: output file name (full path) for the
probability raster (optional)}

\item{scratch_dir}{Charcter: scratch directory}

\item{executable_dir}{Character: directory where the executable is located}

\item{program_name}{Character: name of the executable}
}
\value{
error code
}
\description{
Program Quantiles reads a raster file and computes quantiles
over a moving circular window. For each iteration of the window,
the interquartile range is determined and outliers, based
on a Tukey's fence with k = 1.5, are removed. Quartiles are then
recalculated using the remaining values. Each pixel of the raster with
a value z less than q1 is assigned a value (z-q1)/(q3-q1) and each
pixel with a value greater than q3 is assigned a value of (z-q3)/(q3-q1).
This shows how many interquartile ranges the pixel value is from the first
or third quartile and serves as a measure of how extreme the value is.
Quantiles will also output z scores for each pixel.
}
