% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{quantiles_input}
\alias{quantiles_input}
\title{Create an input file for Fortran program Quantiles.}
\usage{
quantiles_input(
  in_raster = "nofile",
  radius = 0,
  buffer = 0,
  out_outlier = "nofile",
  out_q1 = "nofile",
  out_q2 = "nofile",
  out_q3 = "nofile",
  out_mean = "nofile",
  out_zscore = "nofile",
  out_prob = "nofile",
  scratch_dir = "none"
)
}
\arguments{
\item{in_raster}{Character: The file name (full path) for the input raster.}

\item{radius}{Numeric (dbl): radius in meters for the moving window}

\item{buffer}{Numeric (dbl): spacign between moving window center points,
in raster cells.}

\item{out_outlier}{Character: output file name (full path) for the 
outlier raster (optional)}

\item{out_q1}{Character: output file name (full path) for the
first-quartile raster (optional)}

\item{out_q2}{Character: output file name (full path) for the
median raster (optional)}

\item{out_q3}{Character: output file name (full path) for the
third-quartile raster (optional)}

\item{out_mean}{Character: output file name (full path) for the mean raster}

\item{out_zscore}{Character: output file name (full path) for the 
z-score raster}

\item{out_prob}{Character: output file name (full path) for the
probability raster (optional)}

\item{scratch_dir}{Charcter: scratch directory}
}
\value{
There is no explicit return object, but an explicit side effect
  is writing to disk of the partial input file.
}
\description{
Program Quantiles reads a raster file and computes quantiles
over a moving circular window. For each iteration of the window,
the interquartile range is determined and outliers, based
on a Tukey's fence with k = 1.5, are removed. Quartiles are then
recalculated using the remaining values and the minimum and 
maximum range
}
