% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_metrics.R
\name{resample}
\alias{resample}
\title{Resample}
\usage{
resample(
  in_raster = "nofile",
  skip = 0,
  out_raster = "none",
  scratch_dir = "none"
)
}
\arguments{
\item{in_raster}{Character string: the DEM to be down sampled}

\item{skip}{Integer, the output cell length is skip * input cell length.
So if the input DEM has cell size 1m and skip is 2, the output DEM
has cell size 2m. If skip is 5, the output DEM has cellsize 5m.}

\item{out_raster}{The output DEM, written as a .flt file.}

\item{scratch_dir}{Character string: A scratch directory where temporary
files are written. If an input file for program partial is created,
it is written here.}
}
\value{
A .flt raster written to disk.
}
\description{
Read an input DEM (.flt or .tif) and output a DEM of lower resolution.
}
\details{
Resample operates as a wrapper for program resample.
}
