% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utils.R
\name{resample_input}
\alias{resample_input}
\title{Create an input file for Fortran program resample}
\usage{
resample_input(in_raster, skip, out_raster, scratch_dir)
}
\arguments{
\item{in_raster}{Character string}

\item{skip}{integer, number of cells to skip.}

\item{out_raster}{Character string, resampled input raster}

\item{scratch_dir}{Character string: Directory for temporary files.
The input file is written to the scratch_dir.}
}
\value{
There is no explicit return object, but an explicit side effect
  is writing to disk (in the scratch_dir) of the input file.
}
\description{
Resample downsamples a raster using increments of the raster pixel size.
It uses no interpolation; output pixel corner locations match input pixel corners,
the pixels are just bigger. Resample is primarily used to downsample a DEM,
in which case instead of pixel corners, we're dealing with grid points or cells.
}
