% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRandomForest.R
\name{runRandomForest}
\alias{runRandomForest}
\title{run random forest}
\usage{
runRandomForest(
  workingDir,
  modelFile,
  referenceRasterFile,
  inputRasterFiles,
  inputPolygonFiles,
  testPointsFile,
  classFieldName,
  wetlandClass,
  nonwetlandClass,
  calcStats,
  probRasterName
)
}
\arguments{
\item{workingDir}{Working directory where model files can be found and output files will be saved in}

\item{modelFile}{Filename of the model}

\item{referenceRasterFile}{Raster to use as a grid reference}

\item{inputRasterFiles}{List of input raster filenames}

\item{inputPolygonFiles}{List of input polygon filenames}

\item{testPointsFile}{Filename of point feature classified by wetland type}

\item{classFieldName}{Name of the class field in the test dataset}

\item{wetlandClass}{Class name for wetlands}

\item{nonwetlandClass}{Class name for non-wetlands}

\item{calcStats}{Calculate ROC statistics for the built model?}

\item{probRasterName}{Filename of the generated wetland probability raster}
}
\description{
run random forest
}
