% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_from_polygons}
\alias{sample_from_polygons}
\title{Sample points from polygons}
\usage{
sample_from_polygons(polygons, sampleRate)
}
\arguments{
\item{polygons}{SpatVector object containing polygons}

\item{sampleRate}{Number of points per kilometer to sample}
}
\value{
SpatVector of sampled points
}
\description{
Given a set of polygons, randomly sample points
from within polygons.
}
