% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_training_data.R
\name{sample_negative_points}
\alias{sample_negative_points}
\title{Sample negative points}
\usage{
sample_negative_points(
  positive_points,
  analysis_region,
  buffer = TRUE,
  buffer_radius = 15,
  negative_proportion = 1,
  rseed = NULL
)
}
\arguments{
\item{positive_points}{SpatVector with locations of all points with positive
class}

\item{analysis_region}{SpatRaster with non-NA values everywhere that
points can be sampled from. All locations that should be excluded from
sampling should be NA. If NULL, all cells which are non-NA for all layers
of the predictors_raster will be used.}

\item{buffer}{Return buffer around points?}

\item{buffer_radius}{minimum possible distance between a positive and negative
point}

\item{negative_proportion}{Proportion of negative points to be generated
compared to number of positive points}

\item{rseed}{Optional integer to seed the random sampling. This allows exact
"random"  results to be reproduced multiple times. If no number is given, a
random number will be chosen as a seed.}
}
\value{
SpatVector with positive and negative points, wtih a field
\code{"class"} indicating whether each point is positive or negative
}
\description{
randomly sample points from an analysis area to create a
dataset of negative points given a set of positive points
}
