% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_training_data.R
\name{sample_points}
\alias{sample_points}
\title{Sample Points from a region}
\usage{
sample_points(count, region, buffer = TRUE, radius = 15, rseed = NULL)
}
\arguments{
\item{count}{The number of to sample}

\item{region}{A SpatRaster with NA cells anywhere a sampled point
cannot be located}

\item{buffer}{Should points be returned with a buffer?}

\item{radius}{The radius of each buffer (ignored if buffer = FALSE)}

\item{rseed}{Optional integer to seed the random sampling. This allows exact
"random"  results to be reproduced multiple times. If no number is given, a
random number will be chosen as a seed.}
}
\value{
A SpatVector of buffers.
}
\description{
Generates a SpatVector of points in a given
area.
}
\details{
This extends \code{terra::spatSample()} to generate the correct
number of sampled points when the sample raster has a lot of NAs.
}
