% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingData.R
\name{sampleNegativePoints}
\alias{sampleNegativePoints}
\title{Sample negative points}
\usage{
sampleNegativePoints(
  positivePoints,
  analysisRegion,
  buffer = TRUE,
  bufferRadius = 15,
  negativeProportion = 1
)
}
\arguments{
\item{positivePoints}{SpatVector with locations of all points with positive
class}

\item{analysisRegion}{SpatRaster with non-NA values everywhere that
points can be sampled from. All locations that should be excluded from
sampling should be NA. If NULL, all cells which are non-NA for all layers
of the predictorsRaster will be used.}

\item{buffer}{Return buffer around points?}

\item{bufferRadius}{minimum possible distance between a positive and negative
point}

\item{negativeProportion}{Proportion of negative points to be generated
compared to number of positive points}
}
\value{
SpatVector with positive and negative points, wtih a field
\code{"class"} indicating whether each point is positive or negative
}
\description{
randomly sample points from an analysis area to create a
dataset of negative points given a set of positive points
}
