% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingData.R
\name{samplePoints}
\alias{samplePoints}
\title{Sample Points from a region}
\usage{
samplePoints(count, region, buffer = TRUE, radius = 15)
}
\arguments{
\item{count}{The number of to sample}

\item{region}{A SpatRaster with NA cells anywhere a sampled point
cannot be located}

\item{buffer}{Should points be returned with a buffer?}

\item{radius}{The radius of each buffer (ignored if buffer = FALSE)}
}
\value{
A SpatVector of buffers.
}
\description{
Generates a SpatVector of points in a given
area.
}
\details{
This extends \code{terra::spatSample()} to generate the correct
number of sampled points when the sample raster has a lot of NAs.
}
