% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{samplePolys}
\alias{samplePolys}
\title{Sample points from polygons}
\usage{
samplePolys(polys, sampleRate)
}
\arguments{
\item{polys}{SpatVector object containing polygons}

\item{sampleRate}{Number of points per kilometer to sample}
}
\value{
coordinates at sample points
}
\description{
Given a set of polygons, randomly sample points
from within polygons.
}
