% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerrainWorksUtils-package.R
\docType{package}
\name{TerrainWorksUtils-package}
\alias{TerrainWorksUtils}
\alias{TerrainWorksUtils-package}
\title{TerrainWorksUtils: General utility functions for TerrainWorks}
\description{
TerrainWorks resources for creating and assessing statistical models.
}
\author{
\strong{Maintainer}: Dan Miller \email{dan@terrainworks.com}

Authors:
\itemize{
  \item Helen Miller
  \item Tate Brasel
  \item Julia Lober
}

}
\keyword{internal}
