% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceMetrics_utils.R
\name{write_input_file_MakeGrids}
\alias{write_input_file_MakeGrids}
\title{Write input file for MakeGrids.exe}
\usage{
write_input_file_MakeGrids(
  DEM_path,
  length_scale,
  output_dir = getwd(),
  grad = TRUE,
  plan = TRUE,
  prof = TRUE,
  bcon = TRUE,
  filename = file.path(output_dir, "input_makeGrids.txt"),
  overwrite = TRUE,
  output_file_extension = paste0("_", length_scale)
)
}
\description{
Write input file for MakeGrids.exe
}
