% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRandomForest.R
\name{writeModelStats}
\alias{writeModelStats}
\title{display model stats}
\usage{
writeModelStats(rfModelInfo, trainingDf, modelName = "", plotDir = NULL)
}
\arguments{
\item{rfModelInfo}{list with random forest model and metadata about inputs,
as returned by buildRandomForest()}
}
\description{
display model stats
}
