% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignRasters.R
\name{alignRasters}
\alias{alignRasters}
\title{Align rasters}
\usage{
alignRasters(referenceRaster, inputRasters)
}
\arguments{
\item{referenceRaster}{A \code{SpatRaster} object to use as a guide for
alignment.}

\item{inputRasters}{A list of \code{SpatRaster} objects to be aligned.}
}
\value{
A list of \code{SpatRaster} objects that share the same grid and
projection as \code{referenceRaster}.
}
\description{
Aligns the grid and projection of multiple input rasters with a
reference raster. In input raster will be aligned if it doesn't match the
dimensions, resolution, extent, origin, or CRS projection of the reference
raster.
}
