% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignRasters.R
\name{alignRasters}
\alias{alignRasters}
\title{Align Rasters}
\usage{
alignRasters(referenceRaster = NULL, inputRasters = NULL)
}
\arguments{
\item{referenceRaster}{A \code{SpatRaster} object to be aligned with.}

\item{inputRasters}{A list of \code{SpatRaster} objects to align with the
\code{referenceRaster}.}
}
\value{
A list of \code{SpatRaster} objects that share the same grid and
projection as \code{referenceRaster}.
}
\description{
Aligns the grid and projection of multiple input rasters with a
reference raster. An input raster will be aligned if it doesn't match the
dimensions, resolution, extent, origin, or CRS projection of the reference
raster.
}
\examples{
\donttest{
library(WetlandTools)

referenceRaster <- terra::rast("C:/Work/netmapdata/Puyallup/elev_puy.flt")

inputRasters <- list(
  aligned = terra::rast("C:/Work/netmapdata/Puyallup/dev_50.tif"),
  unaligned = terra::rast("C:/Work/netmapdata/Puyallup/grad_50.tif")
)

alignedRasters <- alignRasters(referenceRaster, inputRasters)
}
}
