% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRocStats.R
\name{calcRocStats}
\alias{calcRocStats}
\title{Calculate ROC Statistics}
\usage{
calcRocStats(classes = NULL, probs = NULL, posClass = NULL, negClass = NULL)
}
\arguments{
\item{classes}{A character vector of classifications.}

\item{probs}{A numeric vector of probability values corresponding to the
elements of \code{classes}.}

\item{posClass}{Name of the positive class.}

\item{negClass}{Name of the negative class.}
}
\value{
A list of ROC statistic objects and values.
}
\description{
Calculates ROC statistics for a set of classes and their
probabilities data
and a wetland probability raster. ROC statistics include: ROC, area under ROC
curve (AUC), precision, accuracy.
}
\examples{
\donttest{
library(TerrainWorksUtils)

# Generate test data with a bit of class probability overlap
classes <- c(rep("WET", 40), rep("UPL", 60))
probs <- c(runif(40, 0.4, 1.0), runif(60, 0.0, 0.6))

rocStats <- calcRocStats(classes, probs, "WET", "UPL")
}

}
