% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceMetrics.R
\name{calculate_surface_metrics}
\alias{calculate_surface_metrics}
\title{Calculate surface metrics}
\usage{
calculate_surface_metrics(
  metrics = c("grad", "plan", "prof", "bcon", "dev", "twi"),
  DEM_path,
  output_dir = getwd(),
  length_scale = 15,
  output_suffix = paste0("_", length_scale),
  dev_resample = 1,
  dev_interval = 1,
  pca_hours = 48,
  pca_conductivity = 1
)
}
\description{
Calculate surface metrics
}
