% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingData.R
\name{createTrainingDataFromPoints}
\alias{createTrainingDataFromPoints}
\title{Create Training Data from Points}
\usage{
createTrainingDataFromPoints(
  positivePoints,
  predictorsRaster,
  analysisRegionMask = NULL,
  bufferRadius = 15,
  negativeProportion = 1,
  extractionMethod = "center",
  extractionLayer = NULL
)
}
\arguments{
\item{positivePoints}{SpatVector with locations of all points with positive
class}

\item{predictorsRaster}{SpatRaster with a layer for each predictor variable}

\item{analysisRegionMask}{SpatRaster with non-NA values everywhere that
points can be sampled from. All locations that should be excluded from
sampling should be NA. If NULL, all cells which are non-NA for all layers
of the predictorsRaster will be used.}

\item{bufferRadius}{minimum possible distance between a positive and negative
point}

\item{negativeProportion}{Proportion of negative points to be generated
compared to number of positive points}

\item{extractionMethod}{Method to use for selecting negative points from
within randomly sampled buffers, of the form \code{c(method, layer)}. \code{
method} can be either "max" or "min", and \code{layer} is the name of one
of the layers in \code{predictorsRaster}.
If not NA, a buffer is drawn around all negative points, and a point is
selected from within the buffer according to extractionMethod. For example:
\code{c("max", "grad_15")} indicates to use the cell with the maximum
value for the layer "grad_15" within the buffer.}
}
\value{
a data.frame with values for positive and negative points
}
\description{
Given an input dataset and set of positive points, generate
randomly sampled negative points and extract predictor values for positive
and negative points.
}
