% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRasterValues.R
\name{extractRasterValues}
\alias{extractRasterValues}
\title{Extract raster values}
\usage{
extractRasterValues(raster = NULL, points = NULL, stringsAsFactors = TRUE)
}
\arguments{
\item{raster}{A \code{SpatRaster} object to extract values from.}

\item{points}{A \code{SpatVector} object of points.}

\item{stringsAsFactors}{Whether to convert character values to factors.}
}
\value{
A \code{data.frame} of values.
}
\description{
Extracts all raster variable values at specified locations.
}
\examples{
\donttest{
library(TerrainWorksUtils)

points <- terra::vect("C:/Work/netmapdata/pack_forest/PF_trainingdata.shp")

# Single-band rasters
rasterContinuous <- terra::rast("C:/Work/netmapdata/pack_forest/pf_dtm3.flt")
rasterFactor <- terra::rast("C:/Work/netmapdata/pack_forest/geo_unit.tif")

v1 <- extractRasterValues(rasterContinuous, points)
v2 <- extractRasterValues(rasterFactor, points)

# Multi-band raster
alignedRasters <- alignRasters(rasterContinuous, list(rasterContinuous, rasterFactor))
rasterStack <- c(alignedRasters[[1]], alignedRasters[[2]])

v3 <- extractRasterValues(rasterStack, points)
}
}
