% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractRasterValues}
\alias{extractRasterValues}
\title{Extract raster values}
\usage{
extractRasterValues(raster = NULL, points = NULL)
}
\arguments{
\item{raster}{A \code{SpatRaster} object to extract values from.}

\item{points}{A \code{SpatVector} object of points.}
}
\value{
A \code{data.frame} of values.
}
\description{
Extracts all raster variable values at specified locations.
}
