% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_training_data.R
\name{make_neg_region}
\alias{make_neg_region}
\title{Make a negative region raster}
\usage{
make_neg_region(
  positive_points,
  ref_raster,
  inner_buffer,
  outer_buffer,
  return_raster = TRUE
)
}
\arguments{
\item{ref_raster}{SpatRaster used as a spatial reference; typically a DEM.}

\item{initiation_points}{A SpatVector that contains the initiation points.}

\item{inner_buffer_rad}{The radius in meters of the inner buffer. This is
the area around the initiation point that should be considered "positive",
that is, within the landslide initiation zone.}

\item{outer_buffer_rad}{The radius in meters of the outer buffer. This is
the area around the landslide that can be considered "negative", that is
where no landslides were observed.}
}
\value{
SpatVector with polygons outlining the presumed negative region.
}
\description{
Creates a negative region with which to sample from that exists
between an inner buffer and an outer buffer around given initiation points.
This method will be used to sample a landslide database that is not
considered a full inventory, i.e., we cannot assume all locations that are
not marked as landslide initiation points did not have landslide.
}
